src/requirements.txt
Django
djangorestframework
djangorestframework-api-key
django-model-utils
django-fsm
django-filter
django-rest-swagger
django-celery-beat==2.2.1
django_cors_headers==3.10.0
django-braces
django-environ

celery
redis
psycopg2-binary==2.8.6
whitenoise
markdown
coreapi
requests
gunicorn
factory-boy
requests-mock

drf-spectacular
drf-spectacular[sidecar]

# additional packages
#tabulate
#boto3
#PyPDF2
#reportlab
#xlsxwriter
#lxml
#python-docx
#shapely
#beautifulsoup4
#msal
#azure-identity
#msgraph-core
